feat(explore-charts): Minor cleanup
